+++ Cmdr/CmdrClient/init.lua
@@ -17,6 +17,7 @@ local Cmdr do
 		ActivationKeys = {[Enum.KeyCode.Semicolon] = true};
 		Enabled = true;
 		MashToEnable = false;
+		ActivationUnlocksMouse = false;
 		PlaceName = "Cmdr";
 		Util = Util;
 		Events = {};
@@ -57,6 +58,34 @@ end
 --- Sets whether or not the console is enabled
 function Cmdr:SetEnabled (enabled)
 	self.Enabled = enabled
+end
+
+--- Sets if activation will free the mouse.
+function Cmdr:SetActivationUnlocksMouse (enabled)
+	self.ActivationUnlocksMouse = enabled
+end
+
+--- Shows Cmdr window
+function Cmdr:Show ()
+	if not self.Enabled then
+		return
+	end
+
+	Interface.Window:Show()
+end
+
+--- Hides Cmdr window
+function Cmdr:Hide ()
+	Interface.Window:Hide()
+end
+
+--- Toggles Cmdr window
+function Cmdr:Toggle ()
+	if not self.Enabled then
+		return self:Hide()
+	end
+
+	Interface.Window:SetVisible(not Interface.Window:IsVisible())
 end
 
 --- Enables the "Mash to open" feature